byogan-api/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
babel=2.6.0=py_1
cudatoolkit=7.5=2
cudnn=5.1=0
docker=4.0.1=pypi_0
flask=1.0.3=pypi_0
flask-cors=3.0.8=pypi_0
flask-sijax=0.4.1=pypi_0
flask-wtf=0.14.2=pypi_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_3
gst-plugins-base=1.14.0=hbbd80ab_1
imageio=2.5.0=py36_0
imagesize=1.1.0=py_0
ipykernel=5.1.0=py36h24bf2e0_1002
ipython=7.3.0=py36h24bf2e0_0
ipython_genutils=0.2.0=py_1
jsonschema=3.0.0a3=py36_1000
libffi=3.2.1=hf484d3e_1005
libgcc-ng=7.3.0=hdf63c60_0
libgfortran-ng=7.2.0=hdf63c60_3
libgpuarray=0.7.6=h14c3975_1003
libiconv=1.15=h14c3975_1004
libopenblas=0.3.3=h5a2b251_3
libpng=1.6.36=h84994c4_1000
libprotobuf=3.6.1=hdbcaa40_1000
libsodium=1.0.16=h14c3975_1001
libstdcxx-ng=7.3.0=hdf63c60_0
libtiff=4.0.10=h648cc4a_1001
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.8=h143f9aa_1005
matplotlib=3.0.2=py36_1002
matplotlib-base=3.0.2=py36h167e16e_1002
mccabe=0.6.1=py_1
mistune=0.8.4=py36h14c3975_1000
mkl=2019.1=144
mkl_fft=1.0.10=py36_0
mkl_random=1.0.2=py36_0
nbconvert=5.3.1=py_1
nbformat=4.4.0=py_1
networkx=2.3=py_0
ninja=1.9.0=h6bb024c_0
notebook=5.7.4=py36_1000
numpy=1.16.0=pypi_0
numpy-base=1.15.4=py36hde5b4d6_0
numpydoc=0.8.0=py_1
packaging=19.0=py_0
pandas=0.24.1=py36hf484d3e_0
pandas-datareader=0.7.0=pypi_0
pandoc=2.6=1
pandocfilters=1.4.2=py_1
pickleshare=0.7.5=py36_1000
pillow=5.4.1=py36h00a061d_1000
pip=19.1.1=pypi_0
pipenv=2018.11.26=pypi_0
plotly=3.6.1=pypi_0
psutil=5.5.1=py36h14c3975_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py36_1000
pycodestyle=2.5.0=py_0
pycparser=2.19=py_0
pydlm=0.1.1.11=pypi_0
pyflakes=2.1.0=py_0
pygments=2.3.1=py_0
pygpu=0.7.6=py36h3010b51_1000
pylint=2.2.2=py36_1000
pymc3=3.6=pypi_0
pyparsing=2.3.1=py_0
pyqt=5.6.0=py36h13b7fb3_1008
pyrsistent=0.14.10=py36h14c3975_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py_0
pytorch=1.0.0=py3.6_cuda9.0.176_cudnn7.4.1_1
pytz=2018.9=py_0
requests=2.18.4=pypi_0
retrying=1.3.3=pypi_0
scikit-image=0.15.0=py36he1b5a44_0
scikit-learn=0.20.2=py36hd81dba3_0
scipy=1.2.1=py36h7c811a0_0
statsmodels=0.9.0=py36h3010b51_1000
testpath=0.4.2=py36_1000
tk=8.6.9=h84994c4_1000
toolz=0.9.0=py_1
torchvision=0.2.1=py_2
typed-ast=1.3.1=py36h14c3975_0
urllib3=1.22=pypi_0
virtualenv=16.6.1=pypi_0
virtualenv-clone=0.5.3=pypi_0
wheel=0.32.3=pypi_0
wrapt=1.11.1=py36h14c3975_0
wtforms=2.2.1=pypi_0
wurlitzer=1.0.2=py36_1000
xlrd=1.2.0=pypi_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.2=h14c3975_1007
yaml=0.1.7=h14c3975_1001
